requirements.txt
opencv-python
face-recognition
numpy
supabase
python-dotenv
Pillow
pandas
openpyxl
